+++ testkeeper/interface/sql_interface.py
@@ -13,6 +13,7 @@
 """
 import os
 from loguru import logger
+import datetime
 from testkeeper.util.sqlalchemy_db_operation import SQLalchemyDbOperation
 from sqlalchemy.orm import sessionmaker
 
@@ -25,6 +26,15 @@ class SqlInterface:
     sqlSession = sql_alchemy.use_connect()
     mul_session = sql_alchemy.use_connect_by_mul_thread()
 
+    def common_update_method(self, table_obj, update_id: str, name: str, value: str):
+        table_obj_instance = self.mul_session.query(table_obj).filter_by(id=update_id).first()
+        logger.info(table_obj_instance.__repr__())
+        if name in table_obj_instance.__dict__:
+            table_obj_instance.__setattr__(name, value)
+            table_obj_instance.updateTime = datetime.datetime.now()
+            self.mul_session.commit()
+        else:
+            raise Exception(f"修改的key:{name} 不存在")
 
 if __name__ == '__main__':
     db_path = os.path.join(os.path.abspath(os.path.dirname(os.path.abspath(os.path.dirname(__file__)))), "testkeeper",
